pyproject.toml
[tool.poetry]
name = "ganesha-fun-world-backend"
version = "0.1.0"
description = ""
authors = ["Henish Patel <200120107182@git.org.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
psycopg2 = "^2.9.9"
requests = "^2.31.0"
gunicorn = "^22.0.0"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
django-debug-toolbar = "^4.3.0"
python-decouple = "^3.8"
setuptools = "^69.5.1"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2024.2"
weasyprint = "^62.0"
redis = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
